Move name-servers module to domains folder
